requirements.txt
flask==1.0.2
pillow==5.1.0
pytesseract==0.2.0
pytest==3.6.0
selenium==3.12.0
